+++ src/main/kotlin/dslpractice/Person.kt
@@ -1,0 +1,3 @@
+package dslpractice
+
+data class Person(val name: String, val company: String?, val skills: Skills)
