+++ src/main/kotlin/dslpractice/Skill.kt
@@ -1,0 +1,3 @@
+package dslpractice
+
+data class Skill(val type: SkillType, val description: String)
