+++ src/main/kotlin/dslpractice/SkillType.kt
@@ -1,0 +1,6 @@
+package dslpractice
+
+enum class SkillType {
+    SOFT,
+    HARD,
+}
