+++ src/main/kotlin/dslpractice/Skills.kt
@@ -1,0 +1,7 @@
+package dslpractice
+
+data class Skills(private val skills: List<Skill> = emptyList()) {
+    fun getSkills(type: SkillType): List<String> {
+        return skills.filter { it.type == type }.map { it.description }
+    }
+}
