+++ src/main/kotlin/dslpractice/SkillsBuilder.kt
@@ -1,0 +1,16 @@
+package dslpractice
+
+class SkillsBuilder {
+    private val skills = mutableListOf<Skill>()
+
+    fun skill(
+        type: SkillType,
+        description: String,
+    ) {
+        skills.add(Skill(type, description))
+    }
+
+    fun build(): Skills {
+        return Skills(skills)
+    }
+}
