+++ src/test/kotlin/dslpractice/DslTest.kt
@@ -42,59 +42,3 @@ class DslTest {
         person.skills.getSkills(SkillType.HARD) shouldBe listOf("Kotlin")
     }
 }
-
-fun introduce(block: PersonBuilder.() -> Unit): Person {
-    return PersonBuilder().apply(block).build()
-}
-
-data class Person(val name: String, val company: String?, val skills: Skills)
-
-class PersonBuilder {
-    private lateinit var name: String
-    private var company: String? = null
-    private var skills: Skills = Skills()
-
-    fun name(value: String) {
-        this.name = value
-    }
-
-    fun company(value: String) {
-        this.company = value
-    }
-
-    fun skills(value: SkillsBuilder.() -> Unit) {
-        this.skills = SkillsBuilder().apply(value).build()
-    }
-
-    fun build(): Person {
-        return Person(name, company, skills)
-    }
-}
-
-data class Skills(
-    private val skills: Map<SkillType, List<String>> = emptyMap(),
-) {
-    fun getSkills(type: SkillType): List<String> {
-        return skills[type].orEmpty()
-    }
-}
-
-class SkillsBuilder {
-    private val skillsMap = mutableMapOf<SkillType, MutableList<String>>()
-
-    fun skill(
-        type: SkillType,
-        description: String,
-    ) {
-        skillsMap.computeIfAbsent(type) { mutableListOf() }.add(description)
-    }
-
-    fun build(): Skills {
-        return Skills(skillsMap.mapValues { it.value.toList() })
-    }
-}
-
-enum class SkillType {
-    SOFT,
-    HARD,
-}
